+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 v1.1.0
+- fix(logging): Improved exception handling to be more robust for HTTP client initialization
+- fix(auth): Implemented .NET native credential negotiation to support both Basic & Digest auth policies
 - chore(docs): Updated doctool release actions to v5
 - chore(build): Updated .NET target frameworks to net8.0 and net10.0
 - fix(odkg): Fixed incorrect mapping of ECDSA algorithm to ECP. Updated to ECDSA.